added options to open countries list in bottom sheet or , dialog
