+++ CountryPicker/src/main/java/com/arpitkatiyarprojects/countrypicker/CountryPickerOutlinedTextField.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
+import com.arpitkatiyarprojects.countrypicker.enums.CountryListDisplayType
 import com.arpitkatiyarprojects.countrypicker.models.BorderThickness
 import com.arpitkatiyarprojects.countrypicker.models.CountriesListDialogDisplayProperties
 import com.arpitkatiyarprojects.countrypicker.models.CountryDetails
@@ -83,6 +84,7 @@ fun CountryPickerOutlinedTextField(
     shape: Shape = RoundedCornerShape(12.dp),
     colors: TextFieldColors = OutlinedTextFieldDefaults.colors(),
     borderThickness: BorderThickness = BorderThickness(),
+    countryListDisplayType: CountryListDisplayType = CountryListDisplayType.Dialog,
     onDone: (() -> Unit)? = null,
 ) {
     PickerOutlinedTextField(
@@ -104,6 +106,7 @@ fun CountryPickerOutlinedTextField(
                 countriesListDialogDisplayProperties = countriesListDialogDisplayProperties,
                 defaultCountryCode = defaultCountryCode,
                 countriesList = countriesList,
+                countryListDisplayType = countryListDisplayType,
                 onCountrySelected = onCountrySelected
             )
         },
