+++ CountryPicker/src/main/java/com/arpitkatiyarprojects/countrypicker/enums/CountryListDisplayType.kt
@@ -1,0 +1,6 @@
+package com.arpitkatiyarprojects.countrypicker.enums
+
+enum class CountryListDisplayType {
+    BottomSheet,
+    Dialog
+}
