+++ app/src/main/java/com/arpitkatiyarprojects/countrypickerproject/ui/country_picker_demo/CountryPickerWithoutTextField.kt
@@ -14,6 +14,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.arpitkatiyarprojects.countrypicker.CountryPicker
+import com.arpitkatiyarprojects.countrypicker.enums.CountryListDisplayType
 import com.arpitkatiyarprojects.countrypicker.models.CountriesListDialogDisplayProperties
 import com.arpitkatiyarprojects.countrypicker.models.CountryDetails
 import com.arpitkatiyarprojects.countrypicker.models.SelectedCountryDisplayProperties
@@ -22,6 +23,7 @@ import com.arpitkatiyarprojects.countrypickerproject.ui.common.CountryDetailsSec
 import com.arpitkatiyarprojects.countrypickerproject.ui.common.SelectedCountrySettings
 import com.arpitkatiyarprojects.countrypickerproject.ui.common.SpacerHeight16
 import com.arpitkatiyarprojects.countrypickerproject.ui.common.SpacerHeight8
+import com.arpitkatiyarprojects.countrypickerproject.ui.common.TextSwitchRow
 import com.arpitkatiyarprojects.countrypickerproject.ui.common.TitleSettingsComposable
 
 @Composable
@@ -39,6 +41,10 @@ fun CountryPickerWithoutTextField() {
         mutableStateOf<CountryDetails?>(null)
     }
 
+    var showCountryListInBottomSheet by remember {
+        mutableStateOf(false)
+    }
+
     Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
         Column(
             modifier = Modifier
@@ -52,6 +58,7 @@ fun CountryPickerWithoutTextField() {
                     .padding(horizontal = 16.dp),
                 selectedCountryDisplayProperties = selectedCountryDisplayProperties,
                 countriesListDialogDisplayProperties = countriesListDialogDisplayProperties,
+                countryListDisplayType = if (showCountryListInBottomSheet) CountryListDisplayType.BottomSheet else CountryListDisplayType.Dialog
             ) {
                 selectedCountryState = it
             }
@@ -66,6 +73,10 @@ fun CountryPickerWithoutTextField() {
                 CountriesListDialogSettings(countriesListDialogDisplayProperties) {
                     countriesListDialogDisplayProperties = it
                 }
+            }
+            SpacerHeight8()
+            TextSwitchRow("Show Country List in BottomSheet", showCountryListInBottomSheet) {
+                showCountryListInBottomSheet = it
             }
         }
     }
