+++ gradle/libs.versions.toml
@@ -13,7 +13,7 @@ libPhoneNumber = "8.13.33"
 kotlinxSerialization = "1.9.0"
 kotlinxSerializationJson = "1.6.3"
 navigationCompose = "2.8.3"
-material3Library = "1.2.1"
+material3Library = "1.3.0"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
